refactor: shift firebase files to utils folder
