+++ src/utils/firebase/firebase-config.js
@@ -1,0 +1,15 @@
+import { initializeApp } from 'firebase/app';
+
+const firebaseConfig = {
+  apiKey: 'AIzaSyCNLGbbH3FTEvV49vNbEpcMRwaYrEB6ApA',
+  authDomain: 'scanlead-17f3a.firebaseapp.com',
+  projectId: 'scanlead-17f3a',
+  storageBucket: 'scanlead-17f3a.appspot.com',
+  messagingSenderId: '292470144917',
+  appId: '1:292470144917:web:d0a3e1b8d2578485cc409c',
+  storageBucket: 'scanlead-17f3a.appspot.com'
+};
+
+const firebaseApp = initializeApp(firebaseConfig);
+
+export default firebaseApp;
